# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface PlayerRegisterInput {
  /**
     * @minLength 2
     * @maxLength 20
     */
  username: string;
  /** @minLength 1 */
  deviceId: string;
}

export type PlayerRank = typeof PlayerRank[keyof typeof PlayerRank];


export const PlayerRank = {
  Bronze: 'Bronze',
  Silver: 'Silver',
  Gold: 'Gold',
  Platinum: 'Platinum',
  Diamond: 'Diamond',
} as const;

export interface Player {
  id: string;
  username: string;
  wins: number;
  losses: number;
  coins: number;
  rank: PlayerRank;
  gamesPlayed: number;
  createdAt: string;
}

export type MatchResult = typeof MatchResult[keyof typeof MatchResult];


export const MatchResult = {
  win: 'win',
  loss: 'loss',
} as const;

export interface Match {
  id: string;
  opponentUsername: string;
  result: MatchResult;
  playerScore: number;
  opponentScore: number;
  createdAt: string;
}