# lib/api-zod/src/generated/types/match.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MatchResult } from './matchResult';

export interface Match {
  id: string;
  opponentUsername: string;
  result: MatchResult;
  playerScore: number;
  opponentScore: number;
  createdAt: Date;
}