# lib/api-zod/src/generated/types/player.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PlayerRank } from './playerRank';

export interface Player {
  id: string;
  username: string;
  wins: number;
  losses: number;
  coins: number;
  rank: PlayerRank;
  gamesPlayed: number;
  createdAt: Date;
}