# lib/api-zod/src/generated/types/playerRegisterInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PlayerRegisterInput {
  /**
     * @minLength 2
     * @maxLength 20
     */
  username: string;
  /** @minLength 1 */
  deviceId: string;
}